move dataloader to a seperate module
