+++ gpt2/dataloader.py
@@ -1,0 +1,42 @@
+import torch
+from torch.utils.data import Sampler, DataLoader
+
+from data.dataset import MidiDataset
+
+
+class CyclicalDataLoader:
+    def __init__(
+        self,
+        dataset: MidiDataset,
+        sampler: Sampler,
+        batch_size: int,
+        pin_memory: bool = False,
+        num_workers: int = 0,
+        device: torch.device = torch.device("cpu"),
+    ):
+        self.dataset = dataset
+        self.batch_size = batch_size
+        self.pin_memory = pin_memory
+        self.num_workers = num_workers
+        self.device = device
+        self.dataloader = DataLoader(
+            dataset=dataset,
+            sampler=sampler,
+            batch_size=self.batch_size,
+            pin_memory=self.pin_memory,
+            num_workers=num_workers,
+        )
+        self.iterator = iter(self.dataloader)
+
+    def get_batch(self):
+        try:
+            batch = next(self.iterator)
+        except StopIteration:
+            self.iterator = iter(self.dataloader)
+            batch = next(self.iterator)
+
+        x = batch["source_token_ids"].to(self.device, non_blocking=True)
+        y = batch["target_token_ids"].to(self.device, non_blocking=True)
+        mask = batch["target_mask"].to(self.device, non_blocking=True)
+        prompt_lengths = batch["prompt_length"]
+        return x, y, mask, prompt_lengths
